requirements.txt
openai
streamlit
SpeechRecognition
gTTS
pygame
pyaudio
python-dotenv
python-docx
pocketsphinx